Building cluster - data and domain
